// src/app/shared/interfaces/interface/shefieldDoorsInterface.ts
export interface IDoors {
    id: any;
    className: string;
    type: string;
    name: string;
    article: string;
    avalible: boolean;
    individualSize:boolean;
    preparingTime: string;
    containment: any;
    decoration: string;
    furniture: any;
    construction: any;
    brand?: string;
    image?: any;
    price?: any;
    description?: any;
}